+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - fix: make open external link scheme case insensive #4201
 - target-electron: make sure log of stdio server is also logged to file
 - improve accessibility a little #4133
+- use authname instead of displayname for vcard filename
 
 
 <a id="1_46_8"></a>